pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video-composite"
version = "1.0.0"
description = "Green-screen removal and compositing toolkit for AI-generated images and video."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "Pillow>=9.0",
    "rembg>=2.0",
]

[project.optional-dependencies]
gpu = ["rembg[gpu]>=2.0"]
full = ["scipy>=1.7", "imageio-ffmpeg"]
dev = ["pytest>=7.0"]

[project.scripts]
ai-video-composite = "ai_video_composite.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
